version1.0.2,update WhereClauseBuilder api

+++ src/main/java/com/github/data/query/specification/Criteria.java
@@ -15,9 +15,9 @@ public interface Criteria<T> {
 
     List<? extends Attribute<T>> getFetchs();
 
-    Integer getOffset();
+    Long getOffset();
 
-    Integer getMaxResults();
+    Long getMaxResults();
 
     LockModeType getLockModeType();
